requirements.txt
attrs==25.3.0
black==25.1.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pytest==8.3.5
requests==2.32.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
